feat: Create base input style to re-use for select
